+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
@@ -97,6 +97,9 @@ abstract class FrameworkExtensionTest extends TestCase
         $this->assertEquals('session.storage.native', (string) $container->getAlias('session.storage'));
         $this->assertEquals('session.handler.native_file', (string) $container->getAlias('session.handler'));
 
+        $this->assertTrue($container->getParameter('session.auto_start'));
+        $this->assertEquals($container->getParameter('session.storage.on_demand_mode'), 'on');
+
         $options = $container->getParameter('session.storage.options');
         $this->assertEquals('_SYMFONY', $options['name']);
         $this->assertEquals(86400, $options['cookie_lifetime']);