+++ src/Symfony/Component/HttpFoundation/Session/Storage/MockFileSessionStorage.php
@@ -30,18 +30,14 @@ class MockFileSessionStorage extends MockArraySessionStorage
     private $savePath;
 
     /**
-     * @var array
-     */
-    private $sessionData;
-
-    /**
      * Constructor.
      *
-     * @param string      $savePath Path of directory to save session files.
-     * @param string      $name     Session name.
-     * @param MetadataBag $metaBag  MetadataBag instance.
+     * @param string      $savePath Path of directory to save session files
+     * @param string      $name     Session name
+     * @param MetadataBag $metaBag  MetadataBag instance
+     * @param integer     $mode     Start on demand mode
      */
-    public function __construct($savePath = null, $name = 'MOCKSESSID', MetadataBag $metaBag = null)
+    public function __construct($savePath = null, $name = 'MOCKSESSID', MetadataBag $metaBag = null, $mode = self::START_ON_DEMAND)
     {
         if (null === $savePath) {
             $savePath = sys_get_temp_dir();
@@ -53,7 +49,7 @@ class MockFileSessionStorage extends MockArraySessionStorage
 
         $this->savePath = $savePath;
 
-        parent::__construct($name, $metaBag);
+        parent::__construct($name, $metaBag, $mode);
     }
 
     /**
@@ -98,7 +94,7 @@ class MockFileSessionStorage extends MockArraySessionStorage
     public function save()
     {
         if (!$this->started) {
-            throw new \RuntimeException("Trying to save a session that was not started yet or was already closed");
+            throw new \RuntimeException("Trying to save a session that was not started yet or was already closed.");
         }
 
         file_put_contents($this->getFilePath(), serialize($this->data));