+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -10,6 +10,8 @@ CHANGELOG
  * added `TimedPhpEngine`
  * added `--clean` option the the `translation:update` command
  * added `http_method_override` option
+ * Reintroduce `auto_start` session config flag to instruct the `SessionListener` to manually start session
+ * Added session config option `on_demand_mode` to control session start on demand.
 
 2.2.0
 -----